// frontend/src/client/types.gen.ts
export type ImageQuizCreate = {
  title: string
  description?: string | null
  questions: Array<ImageQuizQuestionCreate>
}

export type ImageQuizQuestionCreate = {
  x: number
  y: number
  label: string
}